cut content wip
